+++ Package.swift
@@ -16,8 +16,9 @@ let package = Package(
             targets: ["SnabblePhoneAuth"]),
     ],
     dependencies: [
-        .package(name: "SnabbleNetwork", url: "https://github.com/snabble/snabble-network-ios-sdk", branch: "main"),
-    ],
+//        .package(name: "SnabbleNetwork", url: "https://github.com/snabble/snabble-network-ios-sdk", branch: "main"),
+        .package(name: "SnabbleNetwork", url: "https://github.com/snabble/snabble-network-ios-sdk", branch: "APPS-1350"),
+   ],
     targets: [
         .target(
             name: "SnabblePhoneAuth",