+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Change country structure #APPS-1350
 
 ### Updated 
-* snabble/snabble-network-ios-sdk 4074a7b2b1538363a1e9be7175ed7bb80eaac341 (was 1.0.0)
+* snabble/snabble-network-ios-sdk 02e5385545d79294f405e8aa657ff7e231f7c16e (was 1.0.0)
 * apple/swift-crypto.git 3.2.0
 
 ## [1.1.0]